moved config optimizer and lr out of fit function
